Replaced Effects List with the ItemTable component tree and configured to work accordingly.
Added ability to override the create function in the ItemCreateButton component.

Moved addEffect to FoundryAdapter.

Added the ItemImage component to represent item images for non-clickable row icons like in Effects tab. Allow the ItemTableFooter to also receive this create functiona dn forward it to the child create button.

Refactored item-image class to be item-use-button, since item-image was more appropriate for the new component.