Began work on Item Feat.
Fixed a major missing feature of the editor by having activateCoreLIsteners be called against each of the initializing editors. This caused the Reveal/Hide button to be restored to the editor.

Updated all editors to individually rerender on submission and also when their HTML changes, which occurs when the Reveal/Hide button is clicked.