Reviewed current implemented settings and made sure that settings were not being accidentally applied to inappropriate sheets.
Extracted more CSS variables.

Made HP bar transition smoothly when adjusting hit points. Clamped the width of the bar to 0-100.

Added missing height 100% now that token-link-wrapper is in the mix for NPCs.

Fixed the styles for limited NPC view.

Added missing lastSubmissionTime updates for NPCs. This helps the sheet editor to work.