- [x] Figure out how to keep the named color rather than converting it to hexa when the user provides it. Only put the named color to hexa for the color picker to use, allowing the user to riff on a named color
- [x] Extract file processing code to util-oriented location

- [x] Extract theme validation / parsing to theme.ts

Also made a faint outline around the button menu to assist with delineating it form the rest of the content.

Extracted the color setting article to its own component to allow for more control over store management, i.e. prevent the color picker from updating the store when it is not open.