Consolidated `isSheetUnlocked` to its own foundry adapter function and updated all affected callers.
Added guard clause to applyTitleToWindow. Still awaiting v11 testing in a live game world to see if this resolves the issue.

Fixed issue where the TextInput component cannot reselect all content in the input after making changes to the actor.

Updated dev server proxy so that it overrides requests for style.css and provides an mostly empty file instead of requiring manual labor from the developer when wanting to run the dev server.

Added `prepare-dist-for-dev.js` script that is called each time `npm run dev` is invoked, so that the `dist` folder is in the ideal state for doing live-updated development, with no need for an initial build or other manual processes.