Removed well-intentioned but frustrating a11y suggestions. There are some features only available to mouse users (middle-click to edit, hovering to make certain things visible across components), and alternatives exist for keyboard users with context menu and classic controls.
Reworked item use button component so that the button can be revealed when focus-visible or when the cell is hovered. Added as much keyboard support as able for toggling item description (space bar is inexplicably ignored on all keyboard events).

Revamped health percentage so that it only uses HP value and max for overlays. This is to keep with the handling of incapacitated state. 0 HP value means incapacitation, and therefore, it means 0% health / 100% dead overlay.

Generalized the Bar component and refashioned ItemHpBar as just HpBar, which simply applies the HP color to the Bar component. Gutted out actor HP bar and updated it to use the generalized HP Bar. Ditto for Items.

Added `colorPickerHpBar` setting and allowed HP Bar color to be adjusted in Theme Settings.