Resolved svelte-check errors/warnings with - TidyDropdownList.svelte -> InlineTextDropdownList.svelte
Extended button menu to allow a button class to for applying styles to the menu opener.

Revamped renamed the TidyDropdownList to InlineTextDropdownList, which leverages the ButtonMenu under the hood.

Eliminated z-indexes where able. Tagged over locations for future evaluation. Sometimes this involved moving HTML blocks around so that they layer correctly by virtue of their location in the DOM.