[#84] Calculate Calculate-able Roll Terms in Actions Tab (#95)
* Began work to structure the alternate data mapping.

* Took some notes directly in the code while preparing to replace the previous model with the new model that has calculation simplifications.

Swapped out previous model with new model with stub derived data.

* Finished formula simplification by having intermediate roll terms evaluated and put into numeric terms.

Added on-the-rails flavor removal as an option on the simplification function.

* Updated sheets ready hook to use the real Tidy namespace.

Truncated damage type texts that get too long.

* Swapped actions prop over to the new model with calculated derived damage.

Removed v2 stub.

Bumped maximum dnd5e system version to 2.4.9 to account for more patch releases.

* Removed commented code.

Adjusted the order of the action sets to put Other at the end.

* Reordered functions.

* Did some extra work to ensure the flavor text is being removed.

Accounted for empty string formulae like the crossbow bolt. For now, it is returning back an empty string rather than throwing an error.

* Extracted formula simplification and cantrip scaling to a util script for formulae.

Pulled cantrip scaling and label handling into the mapActionItem function and out of the template.

* Added support for not showing "None" in the range column.

Pulled range title and subtitle back into the action item model.

If there is not a range title but there is a range subtitle, the range will default to "—".

Added more column width for Range and Hit/DC.

* Removed unused code.

* Crunching down.

* Removed tool and backpack hard filter from action set construction. If the user wants a backpack in the action list, then so be it!

Hardened handling of range detection.