Figured out one solution to get buttons functioning.
Adjusted classes and attributes to align the stars and get secrets event handling to work.

Leveraged an open event with an async jquery listener wiring for after the content renders on the accordion.

This produces some animation anomalies when opening small accordion items with secrets, because the animation accounts for the original height and not the expanded height with the additional button(s). Will endeavor to get that to work.