Pivoted Sheet Settings to User Settings.
Delegated user setting names and hints to the underlying settings object options instead of using hardcoded copies of loc keys.

Specified the settings properties for UserSettingsContext so that they are specific and hold no surprises. Upfront labor, but the longterm maintenance (and optional refactors) of this will be pretty easygoing.

Trimmed down the user settings form application configuration