Completed the prototype design and general approach for Filter Pins and demonstrated it on the PC inventory tab.
Created PinnedFilterToggle and PinnedFilterToggles components for managing the toggles.

Replaced primary accent color for active/hovered filter menu with the icon hover color, now that there are more colors on the toolbar. It was slightly awkward before, but with the pinned filter toggles, it's just messy looking with an accent color.

Added abbreviation support for pinned filters. Added abbreviation for concentration.