[#478] Default Pinned Filters on Utility Bar (#487)
* Began prototype for default filter pins.

* Completed the prototype design and general approach for Filter Pins and demonstrated it on the PC inventory tab.

Created PinnedFilterToggle and PinnedFilterToggles components for managing the toggles.

Replaced primary accent color for active/hovered filter menu with the icon hover color, now that there are more colors on the toolbar. It was slightly awkward before, but with the pinned filter toggles, it's just messy looking with an accent color.

Added abbreviation support for pinned filters. Added abbreviation for concentration.

* Propagated default pinned filters out to the character sheet in full.

* Propagated pinned filters to NPCs, vehicles, and containers.

Extracted filter styles to a component stylesheet partial, since there are some issues with specificity because of the legacy item sheets.

Reworked the getPinnedFiltersForTab to prevent duplicate entries, just in case.

* Added loc keys for Can Cast and Can Use.

* Added SpellUtils and ItemUtils for determining information about items/spells.

* Added canCastSpell filter and put in the NPC and PC spellbook filters. Also added as default pinned filters to replace Prepared.

Updated "Prepared" filter to reuse spell util logic and to account for cantrip preparation.

Updated castable spells to include specifically unlimited (not configured for limited uses) atwill and innate spells rather just atwill and innate by mode.

* Added canUse filter and applied to character features and as a pinned filter.

* Made legendary panel show/hide animated 💅