Added sheet preference by actor type for spell slot tracker mode: pips vs. value/max.
Removed client setting for spell marker from code and localization.

Wired up pips and value/max slot trackers to be exclusive from each other, based on the preference.

Added SheetPreferencesRuntime service that can expose a store for sheets to subscribe to and to re-render to. Removed manual re-rendering for preferences-based changes in sheet logic.