Added special EffectToggleControl component for dealing with active effect enable/disable. Because concentration is a special replacement of the Disable control, it represents a third state in that particular slot on the classic controls menu.
Added support for Dnd5eIcon in ItemControl, by way of an iconSrc string prop.

Extracted concentration effect detection to a FoundryAdapter function.

Added common context field for special saves / concentration.

Made proficiency and movement configs show/hide when the sheet is unlocked/locked.

Added SpecialSaves component and put it to the right of the ActorMovement component in NPC and Character sheet headers.