Added generic function for getting default sort order of any standard array of sections.
Updated sortKeyedSections to leverage the default sort order when a section does not have a configured sort.

Extracted partitionItem function to CharacterSheetSections.

Began process of moving sorting to after sectioning. Generally, this is going to be the way to handle sorting, because sectioning and partitioning is not always handled in the order in which the full collection of items is currently sorted.

Replaced some magic strings with constants in the character sheet.