Implemented facility occupants context menu.
Consolidate context menu entry types to just the Foundry one.

Loosened delete/duplicate restrictions on context menu options so that they're available when the sheet is locked.

Add limitation to favorites so that items cannot be a Facility. This is a temporary stopgap until Facilities are able to be favorited.

Gave context menu items a larger hitbox and cleaner hover style.