Added styling for disabled facilities.
Added some restrictions for disabled facilities. Non-GMs cannot open menus, execute orders, etc. when a facility is disabled.

Adjusted form field styles to allow form field items be justify-content flex-end. Checkboxes were not looking their best when flex-start.

Adjusted the checkbox background to be light color, so that they're more obvious to users in dark mode.

Corrected progress meter when the facility is disabled, not to show value/max and instead to show just Repair and a percentage.