Updated PillSwitch to show just a pill when disabled, and when showing a pill, it will class the pill as `lighter` when unchecked.
Updated all sidebar pills to show proper wording when disabled.

Updated descriptions to only show the first available description when the user is not the owner of the container. Also updated descriptions to hide the edit button when the user is not the owner.

Updated Container Sheet to be ready for Limited and Observer users.