Added support for debugOnly settings.
Added performance mode setting and translated it to sheet window classes. It is visible for configuration when debug is turned on.

disable-shadows and disable-transitions are now added to all Tidy svelte applications when in performance mode.

Performance mode will kick on if Foundry Core perf mode is set to Low (0) or Tidy is in Debug Mode + Tidy Performance Mode is activated.